Combine into single python file
Builds not working when splitting out into separate python files
