--- src/guard_duty_malware_scan.py
@@ -1,36 +1,0 @@
-import time
-import os
-
-
-scan_complete_tag_key = 'GuardDutyMalwareScanStatus'
-threat_found = 'THREATS_FOUND'
-threat_found_result = 'awsGuardDutyThreatFound'
-await_delay_secs = os.getenv('POLL_MALWARE_SCAN_COMPLETE_AWAIT_SECS', 5)
-
-
-def get_object_tagging(s3_client, bucket, object_key):
-    return s3_client.get_object_tagging(
-        Bucket=f'{bucket}',
-        Key=f'{object_key}'
-    )['TagSet']
-
-
-def poll_guard_duty_scan_complete(s3_client, bucket, object_key):
-    print(f'Polling for GuardDuty scan result: {object_key}')
-    while True:
-        tag_set = get_object_tagging(s3_client, bucket, object_key)
-        for tag in tag_set:
-            tag_key = tag['Key']
-            if tag_key == scan_complete_tag_key:
-                scan_result = tag['Value']
-                print(f'Polling for GuardDuty scan result completed: {object_key}')
-                return scan_result
-        time.sleep(await_delay_secs)
-
-
-def guard_duty_threat_found(s3_client, bucket, object_key):
-    scan_result = poll_guard_duty_scan_complete(s3_client, bucket, object_key)
-    threats = []
-    if scan_result == threat_found:
-        threats = [threat_found_result]
-    return threats
